move to run_embeddings and run_features
