+++ src/br/features/classification.py
@@ -8,7 +8,7 @@ from sklearn.preprocessing import StandardScaler
 from tqdm import tqdm
 
 
-def get_classification_df(all_ret, target_col):
+def get_classification_df(all_ret, target_col, df_feat=None):
     ret_dict5 = {
         "model": [],
         "top_1_acc": [],
@@ -16,8 +16,11 @@ def get_classification_df(all_ret, target_col):
         "top_3_acc": [],
         "cv": [],
     }
+
     for model in tqdm(all_ret["model"].unique(), total=len(all_ret["model"].unique())):
         this_mo = all_ret.loc[all_ret["model"] == model].reset_index(drop=True)
+        if df_feat and target_col not in this_mo.columns:
+            this_mo = this_mo.merge(df_feat, on="CellId")
         k1, k2, k3 = get_classification(this_mo, target_col)
         for i in range(len(k1)):
             ret_dict5["model"].append(model)
