+++ src/br/features/regression.py
@@ -17,7 +17,7 @@ def get_regression_df(all_ret, target_cols, feature_df_path, df_feat=None):
     for target in target_cols:
         for model in tqdm(all_ret["model"].unique(), total=len(all_ret["model"].unique())):
             this_mo = all_ret.loc[all_ret["model"] == model].reset_index(drop=True)
-            if feature_df_path and target not in this_mo.columns:
+            if df_feat and target not in this_mo.columns:
                 this_mo = this_mo.merge(df_feat, on="CellId")
             test_r2, test_mse = get_regression(this_mo, target)
             for i in range(len(test_r2)):
