+++ src/br/models/load_models.py
@@ -14,6 +14,8 @@ def load_model_from_path(dataset, results_path, strict=False, split="val", devic
     model_manifest = pd.read_csv(models["orig_df"])
     model_sizes = []
     all_models = []
+    x_labels = []
+    latent_dims = []
     for j, ckpt_path in enumerate(models["model_checkpoints"]):
         if "model_paths" in models.keys():
             config_path = models["model_paths"][j]
@@ -22,6 +24,8 @@ def load_model_from_path(dataset, results_path, strict=False, split="val", devic
         with open(config_path) as stream:
             config = yaml.safe_load(stream)
         model_conf = config["model"]
+        x_label = model_conf["x_label"]
+        latent_dim = model_conf["latent_dim"]
         model_class = model_conf.pop("_target_")
         model_conf = instantiate(model_conf)
         model_class = _locate(model_class)
@@ -31,8 +35,10 @@ def load_model_from_path(dataset, results_path, strict=False, split="val", devic
             ).eval()
         )
         model_sizes.append(config["model/params/total"])
+        x_labels.append(x_label)
+        latent_dims.append(latent_dim)
 
-    return all_models, models["names"], model_sizes, model_manifest
+    return all_models, models["names"], model_sizes, model_manifest, x_labels, latent_dims
 
 
 def load_model_from_mlflow(dataset, results_path, split="val"):
@@ -42,6 +48,8 @@ def load_model_from_mlflow(dataset, results_path, split="val"):
     model_manifest = pd.read_csv(models["orig_df"])
     model_sizes = []
     all_models = []
+    x_labels = []
+    latent_dims = []
     for i in models["run_ids"]:
         all_models.append(
             load_model_from_checkpoint(
@@ -53,13 +61,22 @@ def load_model_from_mlflow(dataset, results_path, split="val"):
         )
         config = get_config(TRACKING_URI, i, "./tmp")
         model_sizes.append(config["model/params/total"])
+        x_labels.append(config["model"]["x_label"])
+        latent_dims.append(config["model"]["latent_dim"])
 
-    return all_models, models["names"], model_sizes, model_manifest
+    return all_models, models["names"], model_sizes, model_manifest, x_labels, latent_dims
 
 
 def get_data_and_models(dataset_name, batch_size, results_path, debug=False):
     data_list = get_data(dataset_name, batch_size, results_path, debug)
-    all_models, run_names, model_sizes, model_manifest = load_model_from_path(
+    (
+        all_models,
+        run_names,
+        model_sizes,
+        model_manifest,
+        x_labels,
+        latent_dims,
+    ) = load_model_from_path(
         dataset_name, results_path
     )  # default list of models in load_models.py
-    return data_list, all_models, run_names, model_sizes, model_manifest
+    return data_list, all_models, run_names, model_sizes, model_manifest, x_labels, latent_dims
